abstracted out settings page, fixed sort on entries page so most recent entries are at the top
